Change Miller's Rabin probability

+++ rsa.h
@@ -2,7 +2,7 @@
 
 void exp_mod(mpz_t x, mpz_t k, mpz_t n, mpz_t *out);
 void inverse_mod(mpz_t x, mpz_t n, mpz_t *out);
-int miller_rabin_test(mpz_t n, int k);
+int miller_rabin_test(mpz_t n, double k);
 void generate_robust_prime(unsigned long b, mpz_t *p, gmp_randstate_t *r_state);
 int string_to_int(char *s, mpz_t n, mpz_t *out);
 void int_to_string(mpz_t *integers, mpz_t n, char *out);
